refactor(payment): enhance paymentClient with getPurchaseById method and update existing methods for clarity
